+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+- Add multi-select support to filter select on the dashboard
+
 ## [v0.0.0] - 2026-03-14
 
 ### Added